+++ src/main/java/org/entur/lamassu/graphql/subscription/handler/EntitySubscriptionHandler.java
@@ -1,4 +1,22 @@
-package org.entur.lamassu.graphql.subscription;
+/*
+ *
+ *
+ *  * Licensed under the EUPL, Version 1.2 or – as soon they will be approved by
+ *  * the European Commission - subsequent versions of the EUPL (the "Licence");
+ *  * You may not use this work except in compliance with the Licence.
+ *  * You may obtain a copy of the Licence at:
+ *  *
+ *  *   https://joinup.ec.europa.eu/software/page/eupl
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the Licence is distributed on an "AS IS" basis,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the Licence for the specific language governing permissions and
+ *  * limitations under the Licence.
+ *
+ */
+
+package org.entur.lamassu.graphql.subscription.handler;
 
 import java.time.Duration;
 import java.util.List;
@@ -6,7 +24,7 @@ import java.util.function.Predicate;
 import org.entur.lamassu.cache.EntityCache;
 import org.entur.lamassu.cache.EntityListener;
 import org.entur.lamassu.model.entities.Entity;
-import org.entur.lamassu.model.subscription.UpdateType;
+import org.entur.lamassu.graphql.subscription.model.UpdateType;
 import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
