+++ src/main/java/org/entur/lamassu/graphql/subscription/handler/VehicleSubscriptionHandler.java
@@ -1,10 +1,29 @@
-package org.entur.lamassu.graphql.subscription;
+/*
+ *
+ *
+ *  * Licensed under the EUPL, Version 1.2 or – as soon they will be approved by
+ *  * the European Commission - subsequent versions of the EUPL (the "Licence");
+ *  * You may not use this work except in compliance with the Licence.
+ *  * You may obtain a copy of the Licence at:
+ *  *
+ *  *   https://joinup.ec.europa.eu/software/page/eupl
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the Licence is distributed on an "AS IS" basis,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the Licence for the specific language governing permissions and
+ *  * limitations under the Licence.
+ *
+ */
+
+package org.entur.lamassu.graphql.subscription.handler;
 
 import java.util.List;
 import org.entur.lamassu.cache.EntityCache;
+import org.entur.lamassu.graphql.subscription.filter.VehicleUpdateFilter;
 import org.entur.lamassu.model.entities.Vehicle;
-import org.entur.lamassu.model.subscription.UpdateType;
-import org.entur.lamassu.model.subscription.VehicleUpdate;
+import org.entur.lamassu.graphql.subscription.model.UpdateType;
+import org.entur.lamassu.graphql.subscription.model.VehicleUpdate;
 import org.entur.lamassu.service.BoundingBoxQueryParameters;
 import org.entur.lamassu.service.GeoSearchService;
 import org.entur.lamassu.service.RangeQueryParameters;
